+++ src/components/common/Modal/index.ts
@@ -1,0 +1,11 @@
+import ModalContent from '@/components/common/Modal/ModalContent';
+import ModalRoot from '@/components/common/Modal/ModalRoot';
+import ModalToggle from '@/components/common/Modal/ModalToggle';
+import ModalPortal from '@/components/common/Modal/Portal';
+
+export const Modal = {
+  Root: ModalRoot,
+  Content: ModalContent,
+  Toggle: ModalToggle,
+  Portal: ModalPortal,
+};
